Add data generator and Kafka producer scripts
This commit adds a script to generate random records from a given dataset (data_generator.py) and another script to push these records to a Kafka producer (kafka_producer.py). It also includes docker-compose file to setup Kafka and Zookeeper services and a `.gitignore` file. Additionally, a script to ensure services are up and running (wait-for-it.sh) is added to the scripts directory.